refactor player exposure to own class, move a helper to utils
